fix(ExtractMetricGraph): clearer name for metric component
