+++ apps/Conduit-UI/src/components/database/DatabaseDashboard.tsx
@@ -1,7 +1,7 @@
 import { Box, Container, Grid } from '@mui/material';
 
 import React from 'react';
-import ExtractGraph from '../metrics/ExtractGraph';
+import ExtractGraph from '../metrics/ExtractMetricGraph';
 import TotalRequestsByModule from '../metrics/TotalRequestsByModule';
 
 const DatabaseDashboard = () => {
