+++ apps/Conduit-UI/src/components/payments/PaymentsDashboard.tsx
@@ -2,7 +2,7 @@ import { Box } from '@mui/material';
 import React from 'react';
 
 const PaymentsDashboard = () => {
-  return <Box>Placeholder</Box>;
+  return <Box>To be implemented</Box>;
 };
 
 export default PaymentsDashboard;
